# server/swagger/swagger.json
{"swagger":"2.0","info":{"title":"The bluelens API","description":"This API provides a set of endpoints to manage users' followees, music history and recommendations.","version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/bluelens/recommendations/{userID}/{maxCount}":{"get":{"tags":["recommendations"],"summary":"recommend recommendations","description":"Make music recommendations for a user.","operationId":"recommendations#recommend","parameters":[{"name":"maxCount","in":"path","description":"Maximum number of recommendations to be returned to the user.","required":true,"type":"integer"},{"name":"userID","in":"path","description":"ID of the user these recommendations are meant for.","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BluelensRecommendations"}}},"schemes":["http"]}},"/bluelens/user/{followerID}/follows/{followeeID}":{"post":{"tags":["user"],"summary":"follow user","description":"Add a user to another user's followees list.","operationId":"user#follow","parameters":[{"name":"followeeID","in":"path","description":"ID of the followee.","required":true,"type":"string"},{"name":"followerID","in":"path","description":"ID of the follower.","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/bluelens/user/{userID}/listen/{musicID}":{"post":{"tags":["user"],"summary":"listen user","description":"Add a music to a user's history.","operationId":"user#listen","parameters":[{"name":"musicID","in":"path","description":"ID of the music.","required":true,"type":"integer"},{"name":"userID","in":"path","description":"ID of the user.","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download server/swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"BluelensMusicLink":{"title":"Mediatype identifier: application/vnd.bluelens.music+json; view=link","type":"object","properties":{"href":{"type":"string","example":"Velit rerum et et ut."},"id":{"type":"string","example":"Necessitatibus exercitationem."}},"description":"A music resource (link view)","example":{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."},"required":["id","href"]},"BluelensMusicLinkCollection":{"title":"Mediatype identifier: application/vnd.bluelens.music+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/BluelensMusicLink"},"description":"BluelensMusicLinkCollection is the media type for an array of BluelensMusicLink (default view)","example":[{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem.","tags":["Cumque dolorum dolorem voluptas eveniet.","Cumque dolorum dolorem voluptas eveniet.","Cumque dolorum dolorem voluptas eveniet."]},{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem.","tags":["Cumque dolorum dolorem voluptas eveniet.","Cumque dolorum dolorem voluptas eveniet.","Cumque dolorum dolorem voluptas eveniet."]}]},"BluelensRecommendations":{"title":"Mediatype identifier: application/vnd.bluelens.recommendations+json; view=default","type":"object","properties":{"links":{"$ref":"#/definitions/BluelensRecommendationsLinks"},"musicID":{"type":"array","items":{"type":"string","example":"Ratione sed ullam aut."},"example":["Ratione sed ullam aut.","Ratione sed ullam aut.","Ratione sed ullam aut."]}},"description":"A list of recommendations for the specified user (default view)","example":{"links":{"list":[{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."},{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."},{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."}],"user":{"href":"Eum deleniti id.","id":"Deserunt quaerat unde dolorem."}},"musicID":["Ratione sed ullam aut.","Ratione sed ullam aut.","Ratione sed ullam aut."]}},"BluelensRecommendationsLinks":{"title":"BluelensRecommendationsLinks","type":"object","properties":{"list":{"$ref":"#/definitions/BluelensMusicLinkCollection"},"user":{"$ref":"#/definitions/BluelensUserLink"}},"description":"BluelensRecommendationsLinks contains links to related resources of BluelensRecommendations.","example":{"list":[{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."},{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."},{"href":"Velit rerum et et ut.","id":"Necessitatibus exercitationem."}],"user":{"href":"Eum deleniti id.","id":"Deserunt quaerat unde dolorem."}}},"BluelensUserLink":{"title":"Mediatype identifier: application/vnd.bluelens.user+json; view=link","type":"object","properties":{"href":{"type":"string","example":"Eum deleniti id."},"id":{"type":"string","example":"Deserunt quaerat unde dolorem."}},"description":"A user resource (link view)","example":{"href":"Eum deleniti id.","id":"Deserunt quaerat unde dolorem."},"required":["id","href"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"OK":{"description":"OK"}}}